// web/src/api/generated/model/announcementForm.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Campus Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { AnnouncementStatus } from './announcementStatus';
import type { Audience } from './audience';

export interface AnnouncementForm {
  id: string;
  title: string;
  body: string;
  audience: Audience;
  pinned: boolean;
  pinned_until: string | null;
  sms_text: string | null;
  status: AnnouncementStatus;
  publish_at: string;
  updated_at: string;
}