// web/src/api/generated/model/announcementIn.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Campus Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Audience } from './audience';
import type { PublishAction } from './publishAction';

export interface AnnouncementIn {
  /** @maxLength 200 */
  title: string;
  /** @maxLength 10000 */
  body: string;
  audience: Audience;
  pinned?: boolean;
  pinned_until?: string | null;
  sms_text?: string | null;
  action: PublishAction;
  publish_at?: string | null;
}