// web/src/api/generated/model/audienceKind.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Campus Portal API
 * OpenAPI spec version: 0.1.0
 */

export type AudienceKind = typeof AudienceKind[keyof typeof AudienceKind];


export const AudienceKind = {
  everyone: 'everyone',
  students: 'students',
  groups: 'groups',
} as const;