// web/src/api/generated/model/composeOptions.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TCFL Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { ProgrammeOption } from './programmeOption';

export interface ComposeOptions {
  from_label: string;
  programmes: ProgrammeOption[];
  title_fits: number;
  sms_max: number;
}