// web/src/api/generated/model/publishAction.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Campus Portal API
 * OpenAPI spec version: 0.1.0
 */

export type PublishAction = typeof PublishAction[keyof typeof PublishAction];


export const PublishAction = {
  draft: 'draft',
  publish: 'publish',
  schedule: 'schedule',
} as const;